+++ arbitrator/wasm-libraries/host-io/src/lib.rs
@@ -19,7 +19,7 @@ extern "C" {
     pub fn wavm_read_customda_preimage(ptr: *mut u8, offset: usize) -> usize;
     pub fn wavm_read_inbox_message(msg_num: u64, ptr: *mut u8, offset: usize) -> usize;
     pub fn wavm_read_delayed_inbox_message(seq_num: u64, ptr: *mut u8, offset: usize) -> usize;
-    pub fn wavm_validate_certificate(ptr: *const u8, preimage_type: u8) -> u32;
+    pub fn wavm_validate_certificate(ptr: *const u8, preimage_type: u8) -> u8;
 }
 
 #[repr(C, align(256))]
@@ -149,7 +149,7 @@ pub unsafe extern "C" fn wavmio__resolveTypedPreimage(
 
 /// Validates a CustomDA certificate, other preimage types are always valid.
 #[no_mangle]
-pub unsafe extern "C" fn wavmio__validateCertificate(preimage_type: u8, hash_ptr: GuestPtr) -> u32 {
+pub unsafe extern "C" fn wavmio__validateCertificate(preimage_type: u8, hash_ptr: GuestPtr) -> u8 {
     let mut our_buf = MemoryLeaf([0u8; 32]);
     let hash = STATIC_MEM.read_slice(hash_ptr, 32);
     our_buf.copy_from_slice(&hash);
